requirements.txt
py-cord[voice]>=2.4.1
websocket-client>=1.7.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
PyNaCl>=1.5.0